fix(admin): fix blank dashboard and users pages
- AdminJobSummary.progress is a 0..1 REAL in the DB, but was typed as
  Int: JSONDecoder cannot decode a fraction into Int, so the whole
  /api/admin/stats response failed to decode and the overview page
  rendered blank. Type it as Double and show a linear progress bar.
- Overview and Users pages rendered an empty Group until the first
  state change; restructure both to always render the List with
  loading/error/content sections inside (same pattern as BookshelfView),
  so they never appear blank.

+++ ZhiZhou/Models/AdminModels.swift
@@ -38,7 +38,8 @@ struct AdminJobSummary: Codable, Identifiable, Hashable {
     let current: Int
     let total: Int
     let chapterCount: Int
-    let progress: Int
+    /// 抓取进度（0~1 小数，与 DB REAL 列一致；对齐 web 端 job.progress * 100）
+    let progress: Double
     let error: String
     let startedAt: Int64
     let updatedAt: Int64
